Reuben 11/28 additions
It compiles so far, and can be played through to a game over screen. If
you find that any of my code is too ambiguous while you are trying to
complete the load and save functions then please let me know (champlain
email is best)

I plan to add more flavor text. maybe a victory condition or final boss,
or it could remain as an infinite how far can you go kinda' game. The
game display needs to be prettied up. perhaps a loot system to replenish
potions and raise power. and it could probably be balanced better too.

on the coding side of the project:
we have the majority of the concepts covered in the course (no casting,
and no enumerated data-types). The code could be organized better, but I
figure we'll look at it once everything is finished to make sure it
follows the coding standard.

added functions to:
1)  generate new character
2) convert an array of strings to int's that can be used for
calculations. (Is this necessary? I assumed that the load function would
pull everything from the input file as strings, but if you can pull all
the info in as the correct data types to begin with then please correct
me!) not implemented in main yet.
3) combat menu and decision
4)  calculate new stats through combat switch statement
5) enemy stat generator
6) death message, continue or exit game.
7 & 8) the convert functions are set apart as I am uncertain whether
they are really necessary. Let me know how the file i/o goes!

header file reorganized.
int main was expanded. rudimentary game is playable.
libraries removed from project.cpp and functions.cpp